Move DatabaseStatistics component out
