+++ ark_nova_stats/frontend/BUILD.bazel
@@ -112,8 +112,8 @@ jest_test(
         "mocks/fileMock.js",
         "src/DateUtils.js",
         "src/DateUtils.test.js",
-        "src/components/ExampleComponent.js",
-        "src/components/ExampleComponent.test.js",
+        "src/components/DatabaseStatistics.js",
+        "src/components/DatabaseStatistics.test.js",
         "src/components/NavBar.js",
         "src/components/PageContainer.js",
         "src/components/PageTitle.js",
